+++ crates/module-system/module-implementations/sov-evm/tests/integration/runtime.rs
@@ -53,6 +53,12 @@ where
 
     type Input = AuthenticatorInput;
 
+    fn max_authentication_gas(&self) -> <S as ::sov_modules_api::Spec>::Gas {
+        S::Gas::from(::sov_modules_api::macros::config_value!(
+            "MAX_AUTHENTICATION_GAS_PER_TX"
+        ))
+    }
+
     fn authenticate(
         &self,
         tx: &AuthenticatorInput,